+++ nodalia-cards-2.1.2.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.1.3-alpha.1.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.3-alpha.1","contentSha256_12":"102d231c598d","file":"nodalia-cards-2.1.2.js","targetFile":"nodalia-cards-2.1.3-alpha.1.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.1.3-alpha.2.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.3-alpha.2","contentSha256_12":"e54f017c6b2c","file":"nodalia-cards-2.1.2.js","targetFile":"nodalia-cards-2.1.3-alpha.2.js","fallbackFile":"nodalia-cards.js"};}