+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.1.3-alpha.1",
-  "contentSha256_12": "102d231c598d",
+  "pkgVersion": "2.1.3-alpha.2",
+  "contentSha256_12": "e54f017c6b2c",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
-    "nodalia-cards-2.1.1.js",
-    "nodalia-cards-2.1.2.js"
+    "nodalia-cards-2.1.2.js",
+    "nodalia-cards-2.1.3-alpha.1.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.1.3-alpha.1.js",
+  "splitCoreFile": "nodalia-cards-core-2.1.3-alpha.2.js",
   "splitCoreSha256_12": "e8c29dcf68d1",
-  "splitSuiteFile": "nodalia-cards-suite-2.1.3-alpha.1.js",
-  "splitSuiteSha256_12": "835609f58525",
-  "editorFile": "nodalia-cards-editor-2.1.3-alpha.1.js",
+  "splitSuiteFile": "nodalia-cards-suite-2.1.3-alpha.2.js",
+  "splitSuiteSha256_12": "dd113161cc80",
+  "editorFile": "nodalia-cards-editor-2.1.3-alpha.2.js",
   "editorSha256_12": "21563a6a4943"
 };
-export const pkgVersion = "2.1.3-alpha.1";
-export const contentSha256_12 = "102d231c598d";
+export const pkgVersion = "2.1.3-alpha.2";
+export const contentSha256_12 = "e54f017c6b2c";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.1.3-alpha.1.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.1.3-alpha.1.js";
-export const editorFile = "nodalia-cards-editor-2.1.3-alpha.1.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.3-alpha.2.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.3-alpha.2.js";
+export const editorFile = "nodalia-cards-editor-2.1.3-alpha.2.js";